requirements.txt
numpy==1.24.2
Pillow==9.4.0
pye57==0.4.1
tqdm==4.64.1